Split up props for `/classify` routes Each /classify route has slightly different logic for generating page props. Handling them all in a single function leads to a bunch of if statements, which might cause bugs. Having a props function for each route seems to be cleaner.
Add project.defaultWorkflow, which returns either the project's single active workflow or its default workflow.
