+++ packages/app-project/pages/projects/[owner]/[project]/classify/workflow/[workflowID]/subject-set/[subjectSetID]/index.js
@@ -1,1 +1,9 @@
-export { default, getServerSideProps } from '@screens/ClassifyPage'
+import getDefaultPageProps from '@helpers/getDefaultPageProps'
+export { default } from '@screens/ClassifyPage'
+
+export async function getServerSideProps({ params, query, req, res }) {
+  const { props: defaultProps } = await getDefaultPageProps({ params, query, req, res })
+  const { subjectSetID, workflowID } = params
+  const props = { ...defaultProps, subjectSetID, workflowID }
+  return ({ props })
+}
