+++ packages/app-project/src/screens/ClassifyPage/index.js
@@ -1,2 +1,1 @@
 export { default } from './ClassifyPageContainer'
-export { getServerSideProps } from './getServerSideProps'
